requirements_local.txt
feedparser
requests
openai
wikipedia
tiktoken
python-dotenv
git+https://github.com/openai/whisper.git
